Fix suggestion admin notification email sending
Some base templates expects a user present in the context when sending
emails. When notifying admins there is none resulting in a 
VariableDoesNotExist error.

We do not have an actual administrative user as such - only an email
provided by the settings. Instead we create a new empty user object and
pass this to the email context.

This will leave placeholders rendering the user display empty. For our
purposes this is fine.